requirements.txt
openai==1.75.0
anthropic==0.49.0
mss==10.0.0
python-dotenv==1.1.0
schedule==1.2.2
pytz==2025.2
pygetwindow==0.0.9
customtkinter==5.2.2
Pillow==11.2.1
pywin32==310
psutil==7.0.0
